refactor: move thumbnail gen tool under `marimo export` ns
