+++ marimo/_cli/export/commands.py
@@ -8,13 +8,13 @@ from typing import Callable, Literal, Optional
 import click
 
 from marimo._cli.export.cloudflare import create_cloudflare_files
+from marimo._cli.export.thumbnail import thumbnail
 from marimo._cli.help_formatter import ColoredCommand, ColoredGroup
 from marimo._cli.parse_args import parse_args
 from marimo._cli.print import (
     echo,
     green,
 )
-from marimo._cli.tools.thumbnails import thumbnail
 from marimo._cli.utils import prompt_to_overwrite
 from marimo._dependencies.dependencies import DependencyManager
 from marimo._dependencies.errors import ManyModulesNotFoundError
