+++ marimo/_cli/export/thumbnail.py
