--- marimo/_cli/tools/__init__.py
@@ -1,1 +1,0 @@
-# Copyright 2026 Marimo. All rights reserved.
